Move integration tests to AbstractS3SDKV2Tests to avoid duplicated mini cluster creation
