+++ hadoop-ozone/integration-test-s3/src/test/java/org/apache/hadoop/ozone/s3/awssdk/v2/AbstractS3SDKV2Tests.java
@@ -21,10 +21,14 @@ import static org.apache.hadoop.ozone.OzoneConsts.MB;
 import static org.apache.hadoop.ozone.s3.awssdk.S3SDKTestUtils.calculateDigest;
 import static org.apache.hadoop.ozone.s3.awssdk.S3SDKTestUtils.createFile;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static software.amazon.awssdk.core.sync.RequestBody.fromString;
 
+import com.google.common.collect.ImmutableList;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
@@ -62,9 +66,13 @@ import org.apache.hadoop.ozone.s3.S3GatewayService;
 import org.apache.hadoop.ozone.s3.endpoint.S3Owner;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.ozone.test.OzoneTestBase;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestMethodOrder;
+import org.junit.jupiter.api.function.Executable;
 import org.junit.jupiter.api.io.TempDir;
 import software.amazon.awssdk.core.ResponseBytes;
 import software.amazon.awssdk.core.sync.RequestBody;
@@ -77,24 +85,43 @@ import software.amazon.awssdk.http.apache.ApacheHttpClient;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.S3Configuration;
+import software.amazon.awssdk.services.s3.model.AbortMultipartUploadRequest;
+import software.amazon.awssdk.services.s3.model.CompleteMultipartUploadRequest;
 import software.amazon.awssdk.services.s3.model.CompleteMultipartUploadResponse;
 import software.amazon.awssdk.services.s3.model.CompletedMultipartUpload;
 import software.amazon.awssdk.services.s3.model.CompletedPart;
 import software.amazon.awssdk.services.s3.model.CopyObjectRequest;
 import software.amazon.awssdk.services.s3.model.CopyObjectResponse;
+import software.amazon.awssdk.services.s3.model.CreateMultipartUploadRequest;
 import software.amazon.awssdk.services.s3.model.CreateMultipartUploadResponse;
+import software.amazon.awssdk.services.s3.model.Delete;
+import software.amazon.awssdk.services.s3.model.DeleteBucketRequest;
+import software.amazon.awssdk.services.s3.model.DeleteObjectTaggingRequest;
+import software.amazon.awssdk.services.s3.model.DeleteObjectsRequest;
+import software.amazon.awssdk.services.s3.model.GetBucketAclRequest;
 import software.amazon.awssdk.services.s3.model.GetObjectRequest;
 import software.amazon.awssdk.services.s3.model.GetObjectResponse;
+import software.amazon.awssdk.services.s3.model.GetObjectTaggingRequest;
+import software.amazon.awssdk.services.s3.model.HeadBucketRequest;
+import software.amazon.awssdk.services.s3.model.HeadObjectRequest;
 import software.amazon.awssdk.services.s3.model.HeadObjectResponse;
 import software.amazon.awssdk.services.s3.model.ListBucketsResponse;
+import software.amazon.awssdk.services.s3.model.ListMultipartUploadsRequest;
 import software.amazon.awssdk.services.s3.model.ListObjectsRequest;
 import software.amazon.awssdk.services.s3.model.ListObjectsResponse;
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Request;
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Response;
+import software.amazon.awssdk.services.s3.model.ListPartsRequest;
+import software.amazon.awssdk.services.s3.model.ObjectIdentifier;
+import software.amazon.awssdk.services.s3.model.PutBucketAclRequest;
+import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.PutObjectResponse;
+import software.amazon.awssdk.services.s3.model.PutObjectTaggingRequest;
+import software.amazon.awssdk.services.s3.model.S3Exception;
 import software.amazon.awssdk.services.s3.model.S3Object;
 import software.amazon.awssdk.services.s3.model.Tag;
 import software.amazon.awssdk.services.s3.model.Tagging;
+import software.amazon.awssdk.services.s3.model.UploadPartCopyRequest;
 import software.amazon.awssdk.services.s3.model.UploadPartRequest;
 import software.amazon.awssdk.services.s3.model.UploadPartResponse;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
@@ -620,5 +647,513 @@ public abstract class AbstractS3SDKV2Tests extends OzoneTestBase {
         .multipartUpload(CompletedMultipartUpload.builder().parts(completedParts).build()));
     assertEquals(bucketName, compResponse.bucket());
     assertEquals(key, compResponse.key());
+  }
+
+  @Nested
+  @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+  class S3BucketVerificationConditionTests {
+
+    private static final String DEFAULT_BUCKET_NAME = "default-bucket";
+    private String correctOwner;
+    private static final String WRONG_OWNER = "wrong-owner";
+    private static final String BUCKET_VERIFICATION_TEST_KEY = "test-key";
+
+    @BeforeAll
+    public void setup() throws Exception {
+      // create bucket to verify bucket ownership
+      s3Client.createBucket(b -> b.bucket(DEFAULT_BUCKET_NAME));
+      GetBucketAclRequest normalRequest = GetBucketAclRequest.builder().bucket(DEFAULT_BUCKET_NAME).build();
+      correctOwner = s3Client.getBucketAcl(normalRequest).owner().displayName();
+
+      // create objects to verify bucket ownership
+      s3Client.putObject(b -> b.bucket(DEFAULT_BUCKET_NAME).key(BUCKET_VERIFICATION_TEST_KEY), RequestBody.empty());
+    }
+
+    @Nested
+    class BucketEndpointTests {
+
+      @Test
+      public void testGetBucketAcl() {
+        GetBucketAclRequest correctRequest = GetBucketAclRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.getBucketAcl(correctRequest));
+
+        GetBucketAclRequest wrongRequest = GetBucketAclRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.getBucketAcl(wrongRequest));
+      }
+
+      @Test
+      public void testListObject() {
+        ListObjectsRequest correctRequest = ListObjectsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.listObjects(correctRequest));
+
+        ListObjectsRequest wrongRequest = ListObjectsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.listObjects(wrongRequest));
+      }
+
+      @Test
+      public void testListMultipartUploads() {
+        ListMultipartUploadsRequest correctRequest = ListMultipartUploadsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.listMultipartUploads(correctRequest));
+
+        ListMultipartUploadsRequest wrongRequest = ListMultipartUploadsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(
+            () -> s3Client.listMultipartUploads(wrongRequest));
+      }
+
+      @Test
+      public void testPutAcl() {
+        PutBucketAclRequest correctRequest = PutBucketAclRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .grantRead("")
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.putBucketAcl(correctRequest));
+
+        PutBucketAclRequest wrongRequest = PutBucketAclRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .grantRead("")
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.putBucketAcl(wrongRequest));
+      }
+
+      @Test
+      public void testHeadBucket() {
+        HeadBucketRequest correctRequest = HeadBucketRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.headBucket(correctRequest));
+
+        HeadBucketRequest wrongRequest = HeadBucketRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        S3Exception exception = assertThrows(S3Exception.class, () -> s3Client.headBucket(wrongRequest));
+        assertEquals(403, exception.statusCode());
+      }
+
+      @Test
+      public void testDeleteBucket() {
+        s3Client.createBucket(builder -> builder.bucket("for-delete"));
+        String newCorrectOwner = s3Client.getBucketAcl(builder -> builder.bucket("for-delete")).owner().displayName();
+
+        DeleteBucketRequest wrongRequest = DeleteBucketRequest.builder()
+            .bucket("for-delete")
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.deleteBucket(wrongRequest));
+
+        DeleteBucketRequest correctRequest = DeleteBucketRequest.builder()
+            .bucket("for-delete")
+            .expectedBucketOwner(newCorrectOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.deleteBucket(correctRequest));
+      }
+
+      @Test
+      public void testMultiDelete() {
+        DeleteObjectsRequest correctRequest = DeleteObjectsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(correctOwner)
+            .delete(Delete.builder().objects(ObjectIdentifier.builder().key("test").build()).build())
+            .build();
+
+        verifyPassBucketOwnershipVerification(() -> s3Client.deleteObjects(correctRequest));
+
+        DeleteObjectsRequest wrongRequest = DeleteObjectsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(WRONG_OWNER)
+            .delete(Delete.builder().objects(ObjectIdentifier.builder().key("test").build()).build())
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.deleteObjects(wrongRequest));
+      }
+    }
+
+    @Nested
+    class ObjectEndpointTests {
+
+      private static final String TEST_CONTENT = "test-content";
+
+      @Test
+      public void testCreateKey() {
+        String newKey = "create-key";
+
+        PutObjectRequest wrongRequest = PutObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.putObject(wrongRequest, fromString(TEST_CONTENT)));
+
+        PutObjectRequest correctRequest = PutObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.putObject(correctRequest, fromString(TEST_CONTENT)));
+      }
+
+      @Test
+      public void testPutObjectTagging() {
+        List<Tag> tags = ImmutableList.of(
+            Tag.builder().key("env").value("test").build(),
+            Tag.builder().key("project").value("example").build()
+        );
+        PutObjectTaggingRequest wrongRequest = PutObjectTaggingRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .tagging(Tagging.builder().tagSet(tags).build())
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.putObjectTagging(wrongRequest));
+
+        PutObjectTaggingRequest correctRequest = PutObjectTaggingRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .tagging(Tagging.builder().tagSet(tags).build())
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.putObjectTagging(correctRequest));
+      }
+
+      @Test
+      public void testCreateMultipartKey() {
+        CreateMultipartUploadRequest wrongRequest = CreateMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.createMultipartUpload(wrongRequest));
+
+        CreateMultipartUploadRequest correctRequest = CreateMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.createMultipartUpload(correctRequest));
+      }
+
+      @Test
+      public void testCreateMultipartByCopy() {
+        String sourceKey = "test-multipart-by-copy-source-key";
+        String destKey = "test-multipart-by-copy-dest-key";
+
+        s3Client.putObject(b -> b.bucket(DEFAULT_BUCKET_NAME).key(sourceKey), fromString(TEST_CONTENT));
+
+        CreateMultipartUploadResponse initResponse =
+            s3Client.createMultipartUpload(b -> b.bucket(DEFAULT_BUCKET_NAME).key(destKey));
+
+        String uploadId = initResponse.uploadId();
+
+        UploadPartCopyRequest wrongRequest = UploadPartCopyRequest.builder()
+            .sourceBucket(DEFAULT_BUCKET_NAME)
+            .sourceKey(sourceKey)
+            .expectedSourceBucketOwner(WRONG_OWNER)
+            .destinationBucket(DEFAULT_BUCKET_NAME)
+            .destinationKey(destKey)
+            .uploadId(uploadId)
+            .partNumber(1)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.uploadPartCopy(wrongRequest));
+
+        UploadPartCopyRequest correctRequest = UploadPartCopyRequest.builder()
+            .sourceBucket(DEFAULT_BUCKET_NAME)
+            .sourceKey(sourceKey)
+            .expectedSourceBucketOwner(correctOwner)
+            .destinationBucket(DEFAULT_BUCKET_NAME)
+            .destinationKey(destKey)
+            .uploadId(uploadId)
+            .expectedBucketOwner(correctOwner)
+            .partNumber(1)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.uploadPartCopy(correctRequest));
+      }
+
+      @Test
+      public void testCopyObject() {
+        String sourceKey = "test-copy-object-source-key";
+        String destKey = "test-copy-object-dest-key";
+        s3Client.putObject(b -> b.bucket(DEFAULT_BUCKET_NAME).key(sourceKey), fromString("test"));
+
+        CopyObjectRequest wrongRequest = CopyObjectRequest.builder()
+            .sourceBucket(DEFAULT_BUCKET_NAME)
+            .sourceKey(sourceKey)
+            .destinationBucket(DEFAULT_BUCKET_NAME)
+            .destinationKey(destKey)
+            .expectedSourceBucketOwner(WRONG_OWNER)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.copyObject(wrongRequest));
+
+        CopyObjectRequest correctRequest = CopyObjectRequest.builder()
+            .sourceBucket(DEFAULT_BUCKET_NAME)
+            .sourceKey(sourceKey)
+            .destinationBucket(DEFAULT_BUCKET_NAME)
+            .destinationKey(destKey)
+            .expectedSourceBucketOwner(correctOwner)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.copyObject(correctRequest));
+      }
+
+      @Test
+      public void testCreateDirectory() {
+        String newKey = "create-directory-key/";
+
+        PutObjectRequest wrongRequest = PutObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.putObject(wrongRequest, RequestBody.empty()));
+
+        PutObjectRequest correctRequest = PutObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.putObject(correctRequest, RequestBody.empty()));
+      }
+
+      @Test
+      public void testGetKey() {
+        GetObjectRequest correctRequest = GetObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.getObject(correctRequest));
+
+        GetObjectRequest wrongRequest = GetObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.getObject(wrongRequest));
+      }
+
+      @Test
+      public void testGetObjectTagging() {
+        GetObjectTaggingRequest correctRequest = GetObjectTaggingRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.getObjectTagging(correctRequest));
+
+        GetObjectTaggingRequest wrongRequest = GetObjectTaggingRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.getObjectTagging(wrongRequest));
+      }
+
+      @Test
+      public void testListParts() {
+        String newKey = "list-parts-key";
+        CreateMultipartUploadResponse multipartUploadResponse = s3Client.createMultipartUpload(b -> {
+          b.bucket(DEFAULT_BUCKET_NAME)
+              .key(newKey)
+              .build();
+        });
+
+        String uploadId = multipartUploadResponse.uploadId();
+
+        s3Client.uploadPart(
+            UploadPartRequest.builder()
+                .bucket(DEFAULT_BUCKET_NAME)
+                .key(newKey)
+                .uploadId(uploadId)
+                .partNumber(1)
+                .contentLength((long) TEST_CONTENT.getBytes(StandardCharsets.UTF_8).length)
+                .build(), fromString(TEST_CONTENT));
+
+        ListPartsRequest correctRequest = ListPartsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .uploadId(uploadId)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.listParts(correctRequest));
+
+        ListPartsRequest wrongResponse = ListPartsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .uploadId(uploadId)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.listParts(wrongResponse));
+      }
+
+      @Test
+      public void testHeadKey() {
+        HeadObjectRequest correctRequest = HeadObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.headObject(correctRequest));
+
+        HeadObjectRequest wrongRequest = HeadObjectRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        S3Exception exception = assertThrows(S3Exception.class, () -> s3Client.headObject(wrongRequest));
+        assertEquals(403, exception.statusCode());
+      }
+
+      @Test
+      public void testDeleteKey() {
+        String newKey = "delete-key";
+        s3Client.putObject(b -> b.bucket(DEFAULT_BUCKET_NAME).key(newKey), fromString(TEST_CONTENT));
+
+        DeleteObjectsRequest wrongRequest = DeleteObjectsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(WRONG_OWNER)
+            .delete(Delete.builder().objects(ObjectIdentifier.builder().key(newKey).build()).build())
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.deleteObjects(wrongRequest));
+
+        DeleteObjectsRequest correctRequest = DeleteObjectsRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .expectedBucketOwner(correctOwner)
+            .delete(Delete.builder().objects(ObjectIdentifier.builder().key(newKey).build()).build())
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.deleteObjects(correctRequest));
+      }
+
+      @Test
+      public void testDeleteObjectTagging() {
+        DeleteObjectTaggingRequest correctRequest = DeleteObjectTaggingRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.deleteObjectTagging(correctRequest));
+
+        DeleteObjectTaggingRequest wrongRequest = DeleteObjectTaggingRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.deleteObjectTagging(wrongRequest));
+      }
+
+      @Test
+      public void testAbortMultipartUpload() {
+        CreateMultipartUploadResponse multipartUploadResponse =
+            s3Client.createMultipartUpload(b -> b.bucket(DEFAULT_BUCKET_NAME).key(BUCKET_VERIFICATION_TEST_KEY));
+
+        String uploadId = multipartUploadResponse.uploadId();
+
+        AbortMultipartUploadRequest wrongRequest = AbortMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .uploadId(uploadId)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.abortMultipartUpload(wrongRequest));
+
+        AbortMultipartUploadRequest correctRequest = AbortMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(BUCKET_VERIFICATION_TEST_KEY)
+            .uploadId(uploadId)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.abortMultipartUpload(correctRequest));
+      }
+
+      @Test
+      public void testInitMultipartUpload() {
+        String newKey = "init-multipart-upload-key";
+
+        CreateMultipartUploadRequest wrongRequest = CreateMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.createMultipartUpload(wrongRequest));
+
+        CreateMultipartUploadRequest correctRequest = CreateMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.createMultipartUpload(correctRequest));
+      }
+
+      @Test
+      public void testCompleteMultipartUpload() {
+        String newKey = "complete-multipart-upload-key";
+        CreateMultipartUploadResponse multipartUploadResponse =
+            s3Client.createMultipartUpload(b -> b.bucket(DEFAULT_BUCKET_NAME).key(newKey));
+
+        String uploadId = multipartUploadResponse.uploadId();
+
+        UploadPartResponse uploadPartResponse = s3Client.uploadPart(b -> b.bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .uploadId(uploadId)
+            .partNumber(1)
+            .contentLength((long) TEST_CONTENT.getBytes(StandardCharsets.UTF_8).length)
+            .build(), fromString(TEST_CONTENT));
+
+        CompletedMultipartUpload completedUpload = CompletedMultipartUpload.builder()
+            .parts(
+                CompletedPart.builder().partNumber(1).eTag(uploadPartResponse.eTag()).build()
+            ).build();
+
+
+        CompleteMultipartUploadRequest wrongRequest = CompleteMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .uploadId(uploadId)
+            .multipartUpload(completedUpload)
+            .expectedBucketOwner(WRONG_OWNER)
+            .build();
+        verifyBucketOwnershipVerificationAccessDenied(() -> s3Client.completeMultipartUpload(wrongRequest));
+
+        CompleteMultipartUploadRequest correctRequest = CompleteMultipartUploadRequest.builder()
+            .bucket(DEFAULT_BUCKET_NAME)
+            .key(newKey)
+            .uploadId(uploadId)
+            .multipartUpload(completedUpload)
+            .expectedBucketOwner(correctOwner)
+            .build();
+        verifyPassBucketOwnershipVerification(() -> s3Client.completeMultipartUpload(correctRequest));
+      }
+    }
+
+    private void verifyPassBucketOwnershipVerification(Executable function) {
+      assertDoesNotThrow(function);
+    }
+
+    private void verifyBucketOwnershipVerificationAccessDenied(Executable function) {
+      S3Exception exception = assertThrows(S3Exception.class, function);
+      assertEquals(403, exception.statusCode());
+      assertEquals("AccessDenied", exception.awsErrorDetails().errorCode());
+    }
   }
 }
